requirements.txt
aiogram==2.25.2
python-dotenv==0.19.0
aiohttp==3.9.1

